Moving cloud on blue background with static rectangle
Set up basic project structure with folder hierarchy and asset locations and names. Displays a sky blue screen containing a single cloud that is movable only along the y-axis along with a static rectangle of a different shade. Whenever the cloud sprite overlaps with the rectangle, the later changes shades to indicate that a collision has occurred.